+++ nall/array.hpp
@@ -16,11 +16,11 @@ namespace nall {
   template<typename T> class array {
   protected:
     T *pool;
-    unsigned poolsize, buffersize;
+    uint64_t poolsize, buffersize;
 
   public:
-    unsigned size() const { return buffersize; }
-    unsigned capacity() const { return poolsize; }
+    uint64_t size() const { return buffersize; }
+    uint64_t capacity() const { return poolsize; }
 
     void reset() {
       if(pool) free(pool);
@@ -29,7 +29,7 @@ namespace nall {
       buffersize = 0;
     }
 
-    void reserve(unsigned newsize) {
+    void reserve(uint64_t newsize) {
       if(newsize == poolsize) return;
 
       pool = (T*)realloc(pool, newsize * sizeof(T));
@@ -37,12 +37,12 @@ namespace nall {
       buffersize = min(buffersize, newsize);
     }
 
-    void resize(unsigned newsize) {
+    void resize(uint64_t newsize) {
       if(newsize > poolsize) reserve(bit::round(newsize));  //round reserve size up to power of 2
       buffersize = newsize;
     }
 
-    T* get(unsigned minsize = 0) {
+    T* get(uint64_t minsize = 0) {
       if(minsize > buffersize) resize(minsize);
       if(minsize > buffersize) throw "array[] out of bounds";
       return pool;
@@ -52,28 +52,28 @@ namespace nall {
       operator[](buffersize) = data;
     }
 
-    template<typename U> void insert(unsigned index, const U list) {
-      unsigned listsize = container_size(list);
+    template<typename U> void insert(uint64_t index, const U list) {
+      uint64_t listsize = container_size(list);
       resize(buffersize + listsize);
       memmove(pool + index + listsize, pool + index, (buffersize - index) * sizeof(T));
       foreach(item, list) pool[index++] = item;
     }
 
-    void insert(unsigned index, const T item) {
+    void insert(uint64_t index, const T item) {
       insert(index, array<T>(item));
     }
 
-    void remove(unsigned index, unsigned count = 1) {
-      for(unsigned i = index; count + i < buffersize; i++) {
+    void remove(uint64_t index, uint64_t count = 1) {
+      for(uint64_t i = index; count + i < buffersize; i++) {
         pool[i] = pool[count + i];
       }
       if(count + index >= buffersize) resize(index);  //every element >= index was removed
       else resize(buffersize - count);
     }
 
-    optional<unsigned> find(const T data) {
-      for(unsigned i = 0; i < size(); i++) if(pool[i] == data) return optional<unsigned>(true, i);
-      return optional<unsigned>(false, 0);
+    optional<uint64_t> find(const T data) {
+      for(uint64_t i = 0; i < size(); i++) if(pool[i] == data) return optional<uint64_t>(true, i);
+      return optional<uint64_t>(false, 0);
     }
 
     void clear() {
@@ -106,13 +106,13 @@ namespace nall {
     }
 
     //index
-    inline T& operator[](unsigned index) {
+    inline T& operator[](uint64_t index) {
       if(index >= buffersize) resize(index + 1);
       //if(index >= buffersize) throw "array[] out of bounds";
       return pool[index];
     }
 
-    inline const T& operator[](unsigned index) const {
+    inline const T& operator[](uint64_t index) const {
       //if(index >= buffersize) throw "array[] out of bounds";
       return pool[index];
     }
