+++ nall/random.hpp
@@ -3,14 +3,14 @@
 
 namespace nall {
   //pseudo-random number generator
-  inline unsigned prng() {
-    static unsigned n = 0;
+  inline uint64_t prng() {
+    static uint64_t n = 0;
     return n = (n >> 1) ^ (((n & 1) - 1) & 0xedb88320);
   }
 
   struct random_cyclic {
-    unsigned seed;
-    inline unsigned operator()() {
+    uint64_t seed;
+    inline uint64_t operator()() {
       return seed = (seed >> 1) ^ (((seed & 1) - 1) & 0xedb88320);
     }
     random_cyclic() : seed(0) {}
