+++ snes/alt/cpu/cpu.cpp
@@ -16,10 +16,10 @@ namespace SNES {
 #include "mmio.cpp"
 #include "timing.cpp"
 
-void CPU::step(unsigned clocks) {
+void CPU::step(uint64_t clocks) {
   smp.clock -= clocks * (uint64)smp.frequency;
   ppu.clock -= clocks;
-  for(unsigned i = 0; i < coprocessors.size(); i++) {
+  for(uint64_t i = 0; i < coprocessors.size(); i++) {
     Processor &chip = *coprocessors[i];
     chip.clock -= clocks * (uint64)chip.frequency;
   }
@@ -42,7 +42,7 @@ void CPU::synchronize_ppu() {
 }
 
 void CPU::synchronize_coprocessor() {
-  for(unsigned i = 0; i < coprocessors.size(); i++) {
+  for(uint64_t i = 0; i < coprocessors.size(); i++) {
     Processor &chip = *coprocessors[i];
     if(chip.clock < 0) co_switch(chip.thread);
   }
